feat: add DTOs for order creation and response

+++ src/main/resources/db/migration/V1__create_shop_table.sql
@@ -4,11 +4,11 @@ CREATE TABLE shopping.shop (
     id BIGSERIAL PRIMARY KEY,
     user_identifier VARCHAR(100) NOT NULL,
     date TIMESTAMP NOT NULL,
-    total FLOAT NOT NULL
+    total NUMERIC(10, 2) NOT NULL
 );
 
 CREATE TABLE shopping.item (
     shop_id BIGSERIAL REFERENCES shopping.shop(id),
     product_identifier VARCHAR(100) NOT NULL,
-    price FLOAT NOT NULL
+    price NUMERIC(10, 2) NOT NULL
 );
